Add an implementation of the NetPBM suite of image formats in the Image hierarchy.
NetPBM is actually six formats in one, providing an ASCII and binary format of bitmap (1-bit), grayscale pixelmap (8/16-bit), and RGB pixelmap (24/48-bit).

Several methods defined in Image::NetPBM to convert between ASCII and binary may be useful in other classes and have been made public static to aid in any future refactoring.  The methods used to parse sections of a longer buffer are similar to the BDB from BIOMDI and may be removed in favor of a C++ BDB replacement in the future, and are also public static.

[git-p4: depot-paths = "//depot/projects/EVALUATION/Main/common/": change = 3197]