Start of a complete rewrite of the Image test program.
This version reads image data out of a RecordStore (test_data/ImageRS).  A set of pre-defined 3-character file extensions tell each Image test driver which Image class to instantiate.  All metadata that can be parsed is displayed.  If Properties information is stored in the Properties RecordStore (test_data/ImagePropertiesRS), the generated metadata is diffed against the recorded data.  If raw and grayscale raw versions (.raw and .gray.raw in ImageRS) are present, the generated raw data is diffed against the recorded raw data.  This version also writes the generated raw data to disk.

Now builds the image test program for JPEGL and NetPBM.

[git-p4: depot-paths = "//depot/projects/EVALUATION/Main/common/": change = 3231]