Add three LaTeX "macros" to link from the textual portion of the document into the doxygen generated LaTeX \classdoxyref, \namespacedoxyref, \structdoxyref.
During the build, these are replaced with styled \hyperlink macros to the doxygen-generated label.  These macros take an optional second parameter to specify the namespace depth of the caption.  By default, just the classname is displayed (depth == 1), but incrementing the number increments the number of namespaces that will be displayed. 

[git-p4: depot-paths = "//depot/projects/EVALUATION/Main/common/": change = 3551]