Remove OpenSSL warnings under OS X 10.7.
Apple deprecated OpenSSL in 10.7 in favor of their own CommonCrypto implementation, available since 10.4.  Starting in 10.7, this resulting in a lot of warnings when building the framework.  OpenSSL will likely be removed in an upcoming release.  Note that this does not check for 10.4 or greater, only Darwin.

Updated exception list and documentation.

[git-p4: depot-paths = "//depot/projects/EVALUATION/Main/common/": change = 4395]