Make NBIS header directories get searched before others for NBIS sources.
This fixes a (future) bug in which the header file for libjpeg (version 9) is read from MacPorts but the header file from libjpegb in NBIS (version 6b) was desired. This triggers missing symbols based on the conditional compilation in be_image_jpegb.cpp (searching for libjpeg <8).

A better solution will be to use the system-provided libjpeg/libjpeg-turbo. Coming soon.

[git-p4: depot-paths = "//depot/projects/EVALUATION/Main/common/": change = 6556]