Record the reserved bits from minutiae Y-coordiantes in ANSI2004 records.
The most significant bits of the Y coordinate in ANSI 2004 records are reserved and should be set to 0. The framework has correctly set the Y coordinate from the lower 14 bits, but did not save the upper two bits of this 2 byte field, making it impossble to use only the framework to confirm that those bits are set to 0.

[git-p4: depot-paths = "//depot/projects/EVALUATION/Main/common/": change = 6991]