Change Framework::Status for Image callbacks.
Framework::Status had been a ready-made object for evaluation test drivers to use for returning from API methods. It ends up not being used much (at all?) because we are rarely using libbiomeval in evaluation APIs. Instead, repurpose this name for a container to store information libbiomeval operations. Make use of this as the parameter for Image:: callbacks.