Enable compilation with OpenSSL >= 1.1.
The functions for EVP_MD_CTX creation/deletion have changed.
Additionally, the EVP_MD_CTX struct cannot be used directly. For
safety, we can wrap this in a unique_ptr and avoid the explicit deletion
call.

OpenSSL 1.1 provides a backwards-compatible API, but does so with the
pre-processor, which does not appear to be compatible with template
arguments. This requires us to branch on the OpenSSL version to
determine the correct functions.