Add capability for RecordStore::mergeRecordStores() to be interrupted.
An optional std::function<> can be passed into mergeRecordStores(). This
function returns a boolean value of true when mergeRecordStores() should
return prematurely. The default std::function<> parameter returns false.
mergeRecordStores() will complete the merging of the current record, but
there is no guarantee that an equal number of records from the source stores
will be merged into the destination.