Finalize support for Release and Debug builds.
Windows must differentiate between both static/dynamic and release/debug. Previously, we built CMake object libraries that would build one object per source file and be used for inclusion in both the static and shared libraries. In Windows, this is not possible. To fix, under Windows, we duplicate the object library targets and have a different build for the dynamic and static versions.

To support both release and debug libraries, we have to change the name. The Windows convention is to postfi the name of the library with "d"