Resolve thread mutex locking problem in Process::Statistics.
The change to the Statistics class that allowed for the ID of the logging task
to be captured requires inter-thread synchronization between the logger thread
and the parent, and now between the parent and the logger to have the logger    wait until the parent has captured the logging thread ID. This can result in
deadlock if the parent, which owns the lock, frees it before the logger
resumes.
Therefore, move the lock and all other inter-thread data to a static object
that will not be deleted.