Don't build video support with force static deps
FFMPEG has many dependencies, and those dependencies are dynamic based
on features built into FFMPEG. A number of those dependencies might
not be built and installed statically on system, so they won't be able
to be found. biomeval doesn't need these dependencies for its own code
but they need to be present or symbols will be missing from FFMPEG
libs.