Create placeholder ID if MPI is not initialized.
Due to error (by programmer, MPI, etc.), it is possible that MPI would
not be initialized by the time generateUniqueID() is called. This
method is often called as a result of throwing an MPI::Exception or
trying to log information about a process, and that info would
subsequently be lost had MPI not be initialized.