Complete migration of framework Video support to current FFMPEG APIs.
Both Video::Container and Video::Stream use the new API.
For Container, changes are to match name changes.
For Stream, packets are sent to the decoder and frames are received back,
matching name changes, and improving he gathering of cached decoded frames.

Improve error handling and message when frames cannot be obtained.
Add three video files with frames that state the frame number. Video frame
rates are 1,10, and 30 frames per second and are useful for testing frame
retrieval.
Change the CMake build settings for the test programs to always turn on the
debug option.