Do not throw from a separate thread.
Because there's no exception handling here, throwing from this thread
will crash. Instead, on error conditions (likely, when a process with
threads that is being monitored starts joining its threads as we are
observing it), break out of the loop immediately and return whatever
information we've gathered so far.

Future implementations may use exception_ptr to communicate the error,
if it is important. However, since the point of this method is to log
information about active threads, and the threads are likely closing,
is it an error or merely the current state of affairs?