Minor code cleanup for the AutoLogger and client classes.
Move log entry creation functions for Process::Statistics to private,
renaming them to be more descriptive, and remove function declaration
left over from previous version.

Remove functions in AutoLogger that are not meant to be part of that class.

In System::MemoryLogger and Process::Statistics improve the descriptions
of the get/setComment functions to describe what they are used for.

Remove calls to deprecated Process::Statistics startAutoLogging()
function in the Process::Statistics test program.