Change how the auto logging comment string is managed, test code improvements.
In the AutoLogger class, manage the comment string for log entries and
have the client classes call the AutoLogger get/set methods. AutoLogger
now adds the comment to each entry.

Improve documentation for the AutoLogger, Process::Statistics, and
MemoryLogger classes,

Change the CMake build for the test programs to use C++20.

Move the rapid-fire test of AutoLogger to the AutoLogger test program,
out of the Statistics test program. Improve testing of the Statistics
class as well.